refactor card sections registry
