+++ src/helpers/cardBuilder.js
@@ -1,13 +1,9 @@
 import { getFlagUrl } from "./country/codes.js";
-import { generateCardTopBar, createNoDataContainer } from "./cardTopBar.js";
 import { safeGenerate } from "./errorUtils.js";
 import { getMissingJudokaFields, hasRequiredJudokaFields } from "./judokaValidation.js";
 import { enableCardFlip } from "./cardFlip.js";
-import {
-  createPortraitSection,
-  createStatsSection,
-  createSignatureMoveSection
-} from "./cardSections.js";
+import { cardSectionRegistry } from "./cardSections.js";
+import { createInspectorPanel } from "./inspector/createInspectorPanel.js";
 
 /**
  * Generates the "last updated" HTML for a judoka card.
@@ -36,126 +32,17 @@ import {
 // }
 
 /**
- * Generates the complete DOM structure for a judoka card.
+ * Build and return a DOM container for a judoka card.
  *
  * @pseudocode
- * 1. Validate the `judoka` object:
- *    - Ensure all required fields are present using `hasRequiredJudokaFields`.
- *
- * 2. Generate the flag URL:
- *    - Call `safeGenerate` with `getFlagUrl` and the `countryCode`.
- *    - Default to "vu" if `countryCode` is missing.
- *    - Fallback to the Vanuatu flag when an error occurs.
- *
- * 3. Determine the card type:
- *    - Use the `rarity` field to set the card type (e.g., "common").
- *
- * 4. Create the main card container:
- *    - Initialize a `<div>` with the class `card-container`.
- *    - Create a child `<div>` with the class `judoka-card` and card type.
- *
- * 5. Add gender-specific styling:
- *    - Add a class based on the `gender` field ("female-card" or "male-card").
- *
- * 6. Append the top bar:
- *    - Generate the top bar using `generateCardTopBar` and append it.
- *    - If generation fails, append a "No data available" container instead.
- *
- * 7. Append the portrait section:
- *    - Generate portrait HTML using `generateCardPortrait`.
- *    - If generation fails, append a "No data available" container.
- *    - Add weight class information to the portrait section.
- *
- * 8. Append the stats section:
- *    - Build the stats panel using `createStatsPanel` and append it.
- *    - If generation fails, append an empty stats container.
- *
- * 9. Append the signature move section:
- *    - Generate signature move HTML using `generateCardSignatureMove` and append it.
- *    - If generation fails, append an empty signature move container.
- *
- * 10. Enable card interactivity:
- *    - Attach click and keyboard handlers to toggle `.show-card-back`.
- *
- * 11. Return the complete card container:
- *    - Append the `judoka-card` to the `card-container`.
- *
- * @param {Object} judoka - The judoka object containing data for the card.
- * @param {Object} gokyo - The Gokyo data (technique information).
- * @returns {HTMLElement} The DOM element for the complete judoka card.
- */
-async function createTopBar(judoka, flagUrl) {
-  return await safeGenerate(
-    () => generateCardTopBar(judoka, flagUrl),
-    "Failed to generate top bar:",
-    createNoDataContainer()
-  );
-}
-
-function createInspectorPanel(container, judoka) {
-  let json;
-  try {
-    json = JSON.stringify(judoka, null, 2);
-  } catch {
-    const p = document.createElement("p");
-    p.textContent = "Invalid card data";
-    return p;
-  }
-
-  const panel = document.createElement("details");
-  panel.className = "debug-panel";
-  panel.setAttribute("aria-label", "Inspector panel");
-
-  const summary = document.createElement("summary");
-  summary.textContent = "Card Inspector";
-  summary.tabIndex = 0;
-  summary.style.minHeight = "44px";
-  summary.style.minWidth = "44px";
-  summary.style.display = "flex";
-  summary.style.alignItems = "center";
-  summary.style.outline = "2px solid transparent";
-  summary.style.outlineOffset = "2px";
-  summary.addEventListener("focus", () => {
-    summary.style.outlineColor = "#000";
-  });
-  summary.addEventListener("blur", () => {
-    summary.style.outlineColor = "transparent";
-  });
-  summary.addEventListener("keydown", (e) => {
-    if (e.key === "Enter" || e.key === " ") {
-      e.preventDefault();
-      panel.open = !panel.open;
-      panel.dispatchEvent(new Event("toggle"));
-    }
-  });
-  panel.appendChild(summary);
-
-  const jsonPre = document.createElement("pre");
-  jsonPre.textContent = json;
-  panel.appendChild(jsonPre);
-
-  // Only show the card's JSON data. The markup preview was removed to
-  // keep the inspector output concise.
-
-  function updateDataset() {
-    summary.setAttribute("aria-expanded", panel.open ? "true" : "false");
-    if (panel.open) {
-      container.dataset.inspector = "true";
-    } else {
-      container.removeAttribute("data-inspector");
-    }
-  }
-
-  panel.addEventListener("toggle", updateDataset);
-  updateDataset();
-
-  return panel;
-}
-
-export { createInspectorPanel };
-
-/**
- * Build and return a DOM container for a judoka card.
+ * 1. Validate `judoka` and `gokyoLookup`:
+ *    - Throw when required fields are missing.
+ * 2. Resolve the flag URL via `safeGenerate(getFlagUrl)`.
+ * 3. Create container and inner card elements and apply gender styling.
+ * 4. For each builder in `cardSectionRegistry`, generate and append the section.
+ * 5. Enable flip interactivity with `enableCardFlip`.
+ * 6. Append an inspector panel when `enableInspector` is true.
+ * 7. Return the populated container.
  *
  * Validates the judoka data and gokyo lookup, generates DOM sections for the card, and assembles them into a single container. When either argument is missing, a fallback element labeled "No data available" is returned.
  *
@@ -203,17 +90,10 @@ export async function generateJudokaCardHTML(judoka, gokyoLookup, options = {}) 
   const genderClass = judoka.gender === "female" ? "female-card" : "male-card";
   judokaCard.classList.add(genderClass);
 
-  const topBarElement = await createTopBar(judoka, flagUrl);
-  judokaCard.appendChild(topBarElement);
-
-  const portraitElement = createPortraitSection(judoka);
-  judokaCard.appendChild(portraitElement);
-
-  const statsElement = await createStatsSection(judoka, cardType);
-  judokaCard.appendChild(statsElement);
-
-  const signatureMoveElement = createSignatureMoveSection(judoka, gokyoLookup, cardType);
-  judokaCard.appendChild(signatureMoveElement);
+  for (const buildSection of cardSectionRegistry) {
+    const section = await buildSection(judoka, { flagUrl, gokyoLookup, cardType });
+    judokaCard.appendChild(section);
+  }
 
   enableCardFlip(judokaCard);
